pyproject.toml
[project]
name = "insurance-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docxtpl>=0.20.2",
    "faker>=40.1.0",
    "flask>=3.1.2",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
